requirements.txt
black==23.9.1
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
exceptiongroup==1.1.3
filelock==3.12.4
fonttools==4.42.1
importlib-resources==6.0.1
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.0
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sympy==1.12
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
zipp==3.16.2
